msg/async/AsyncConnection: always fill the `recv_buf`
Use `readv()` to read into both the caller-provided buffer and the
`recv_buf`, which comes with virtually no extra cost compared to just
filling the caller-provided buffer.  This eliminates lots of system
calls because many single `recv()` calls are combined into one
`readmsg()`.

As a side effect, we can remove the kludge that allocated twice the
`recv_max_prefetch` size.

Signed-off-by: Max Kellermann <max.kellermann@ionos.com>