msg/async/PosixStack: always use recv() instead of read()
recv() works on non-Windows, too, and is the proper system call for
reading from a socket.  It makes no sense to implement different code
paths for Windows and non-Windows.

The distinction was introduced by commit ba35348bee62, but only
because read() was used in all cases previously, which however doesn't
work on Windows, because sockets and file descriptors are different
things on Windows.

Signed-off-by: Max Kellermann <max.kellermann@ionos.com>