mgr/PyModule: do not lock in get_name(), get_error_string()
The way these methods used the lock is completely wrong: in the
protected code block, they do not access anything that needs to be
protected.  All they effectively do is return the `this` pointer with
an offset to the `module_name` or `error_string` field; but they do
not actually access these fields.  The actual access is done by the
caller, after the mutex has been unlocked already.

But anyway, these fields do not need any protection.  Both fields are
`const` and thus cannot be modified by other threads.  So instead of
fixing the return value (removing the `&`), let's just remove the
lock.

Signed-off-by: Max Kellermann <max.kellermann@ionos.com>